requirements.txt
# Core Django
Django==5.0.2
djangorestframework==3.14.0
django-environ==0.11.2
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Database
psycopg2-binary==2.9.9
django-filter==23.5

# API Documentation
drf-spectacular==0.27.1

# File Storage
boto3==1.34.34
django-storages==1.14.2

# Task Queue
celery==5.3.6
redis==5.0.1
django-celery-beat==2.6.0

# Admin Enhancement
django-jazzmin==2.6.0

# Utilities
python-dotenv==1.0.1
Pillow==10.2.0
phonenumbers==8.13.28

# Development
pytest==8.0.0
pytest-django==4.8.0
pytest-cov==4.1.0

# Additional from main branch
channels
channels-redis
gunicorn
coverage
black
flake8
